+++ Http/Controllers/Front/GymTrainingPlanFrontController.php
@@ -23,7 +23,7 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
     {
         $title = trans('sw.training_plans');
         
-        $query = GymTrainingPlan::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1);
+        $query = GymTrainingPlan::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1);
 
         // Search filter
         if ($request->has('q') && $request->q) {
@@ -40,7 +40,7 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
         }
 
         $plans = $query->latest()->paginate(20)->appends($request->except('page'));
-        $total = GymTrainingPlan::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->count();
+        $total = GymTrainingPlan::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->count();
 
         return view('software::Front.training_plan_list', compact('title', 'plans', 'total'));
     }
@@ -52,7 +52,7 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
     {
         $title = trans('sw.add_training_plan');
         $plan = new GymTrainingPlan();
-        $categories = GymSubscriptionCategory::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->get();
+        $categories = GymSubscriptionCategory::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->get();
         
         return view('software::Front.training_plan_form', compact('title', 'plan', 'categories'));
     }
@@ -115,10 +115,10 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
     public function edit($id)
     {
         $title = trans('sw.edit_training_plan');
-        $plan = GymTrainingPlan::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
+        $plan = GymTrainingPlan::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
             ->findOrFail($id);
-        $categories = GymSubscriptionCategory::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->get();
-        $tasks = GymTrainingTask::where('plan_id', $id)->orderBy('order')->get();
+        $categories = GymSubscriptionCategory::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)->get();
+        $tasks = GymTrainingTask::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('plan_id', $id)->orderBy('order')->get();
         
         return view('software::Front.training_plan_form', compact('title', 'plan', 'categories', 'tasks'));
     }
@@ -128,7 +128,7 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
      */
     public function update(GymTrainingPlanRequest $request, $id)
     {
-        $plan = GymTrainingPlan::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
+        $plan = GymTrainingPlan::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
             ->findOrFail($id);
         
         $inputs = $request->all();
@@ -137,7 +137,7 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
         // Update tasks
         if ($request->has('tasks') && is_array($request->tasks)) {
             // Delete old tasks
-            GymTrainingTask::where('plan_id', $id)->delete();
+            GymTrainingTask::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('plan_id', $id)->delete();
             
             // Create new tasks
             foreach ($request->tasks as $taskData) {
@@ -184,13 +184,13 @@ class GymTrainingPlanFrontController extends GymGenericFrontController
      */
     public function destroy($id)
     {
-        $plan = GymTrainingPlan::where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
+        $plan = GymTrainingPlan::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('branch_setting_id', $this->user_sw->branch_setting_id ?? 1)
             ->findOrFail($id);
         
         $name = $plan->title;
         
         // Delete associated tasks
-        GymTrainingTask::where('plan_id', $id)->delete();
+        GymTrainingTask::branch($this->user_sw->branch_setting_id, @$this->user_sw->tenant_id)->where('plan_id', $id)->delete();
         
         $plan->delete();
